pyproject.toml
[project]
name = "caqes-core"
version = "0.0.1"
description = ""
authors = [
    {name = "ADHL",email = "hanlim2002@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "rule-engine (>=4.5.3,<5.0.0)",
]

[tool.poetry]
packages = [{include = "caqes_core", from = "src"}]

[tool.poetry.scripts]
start = "caqes_core.app:main"

[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.25.3"
pytest-httpserver = "^1.1.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
